+++ app/Http/Controllers/EventsController.php
@@ -13,6 +13,7 @@ use App\Models\{
     EventNature,
     Level,
     FundSource,
+    Organization,
 };
 
 use App\Http\Requests\EventRequests\{
@@ -26,18 +27,37 @@ use App\Services\EventServices\{
     EventStoreService,
     EventUpdateService,
     EventDeleteService,
+    EventGetOrganizationIDService,
 };
+use App\Services\PermissionServices\PermissionCheckingService;
 
 class EventsController extends Controller
 {
+    protected $permissionChecker;
+    /**
+     * Create a new controller instance.
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth');
+        $this->permissionChecker = new PermissionCheckingService();
+    }
+
     /**
      * Shows the Index Page of all Events
      * @return View
      */
     public function index()
     {
-        $orgAcronym = Auth::user()->course->organization->organization_acronym;
-        $events = (new EventIndexService())->index();
+        abort_if(! $this->permissionChecker->checkIfPermissionAllows('AR-View_Event'), 403);
+        $orgAcronym = Organization::where('organization_id', (new EventGetOrganizationIDService)->getOrganizationID())->value('organization_acronym');
+        $events = Event::with('eventRole',
+                'eventCategory',
+                'eventLevel',)
+            ->where('organization_id', (new EventGetOrganizationIDService)->getOrganizationID(),)
+            ->orderByRaw('MONTH(`start_date`) ASC, `start_date` ASC')
+            ->paginate(30);
         return view('events.index', compact('events', 'orgAcronym'));
     }
 
@@ -48,6 +68,7 @@ class EventsController extends Controller
      */
     public function show($event_slug, $newEvent = false)
     {
+        abort_if(! $this->permissionChecker->checkIfPermissionAllows('AR-View_Event'), 403);
         abort_if(! Event::where('slug', $event_slug)->exists(), 404);
 
         $event = (new EventShowService())->show($event_slug);
@@ -61,6 +82,7 @@ class EventsController extends Controller
      */
     public function edit($event_slug)
     {
+        abort_if(! $this->permissionChecker->checkIfPermissionAllows('AR-Edit_Event'), 403);
         abort_if(! Event::where('slug', $event_slug)->exists(), 404);
 
         $event = Event::with(
@@ -97,6 +119,7 @@ class EventsController extends Controller
      */
     public function update(EventUpdateRequest $request, $event_slug)
     {
+        abort_if(! $this->permissionChecker->checkIfPermissionAllows('AR-Edit_Event'), 403);
         abort_if(! Event::where('slug', $event_slug)->exists(), 404);
         
         $returnArray = (new EventUpdateService())->update($request, $event_slug);
@@ -120,6 +143,7 @@ class EventsController extends Controller
      */ 
     public function destroy($event_slug)
     {
+        abort_if(! $this->permissionChecker->checkIfPermissionAllows('AR-Delete_Event'), 403);
         abort_if(! Event::where('slug', $event_slug)->exists(), 404);
 
         $message = (new EventDeleteService())->destroy($event_slug);
@@ -135,6 +159,7 @@ class EventsController extends Controller
      */ 
     public function create()
     {
+        abort_if(! $this->permissionChecker->checkIfPermissionAllows('AR-Create_Event'), 403);
         $eventCategories = EventCategory::all();
         $eventClassifications = EventClassification::all();
         $eventNatures = EventNature::all();
@@ -158,6 +183,7 @@ class EventsController extends Controller
      */
     public function store(EventStoreRequest $request)
     {
+        abort_if(! $this->permissionChecker->checkIfPermissionAllows('AR-Create_Event'), 403);
         $returnArray = (new EventStoreService())->store($request);
         $message = $returnArray['message'];
